backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.2
sentence-transformers==2.2.2
chromadb==0.4.18
openai==1.3.5
google-generativeai==0.3.1
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
python-magic==0.4.27
langdetect==1.0.9 